完成涨跌及幅度和买1提示的功能

+++ Mystock-master/app/src/main/java/com/example/rui/mystock/SettingsActivity.java
@@ -59,7 +59,7 @@ public class SettingsActivity extends AppCompatActivity {
             priceFall.setText(String.valueOf(cursor.getDouble(cursor.getColumnIndex("Fall"))));
             priceRiseAmount.setText(String.valueOf(cursor.getDouble(cursor.getColumnIndex("RiseAmount"))));
             priceFallAmount.setText(String.valueOf(cursor.getDouble(cursor.getColumnIndex("FallAmount"))));
-            buy1Set.setText(String.valueOf(cursor.getInt(cursor.getColumnIndex("Buy1Value"))));
+            buy1Set.setText(String.valueOf(cursor.getLong(cursor.getColumnIndex("Buy1Value"))));
 
             if (cursor.getInt(cursor.getColumnIndex("RiseSwitch")) == 1) {
                 priceRiseSwitch.setChecked(true);
